Upgrade to Pulumi v2.12.0 and pulumi-terraform-bridge v2.11.0

+++ sdk/python/pulumi_aiven/kafka_topic.py
@@ -5,7 +5,7 @@
 import warnings
 import pulumi
 import pulumi.runtime
-from typing import Any, Dict, List, Mapping, Optional, Tuple, Union
+from typing import Any, Mapping, Optional, Sequence, Union
 from . import _utilities, _tables
 
 __all__ = ['KafkaTopic']
@@ -16,12 +16,12 @@ class KafkaTopic(pulumi.CustomResource):
                  resource_name: str,
                  opts: Optional[pulumi.ResourceOptions] = None,
                  cleanup_policy: Optional[pulumi.Input[str]] = None,
-                 minimum_in_sync_replicas: Optional[pulumi.Input[float]] = None,
-                 partitions: Optional[pulumi.Input[float]] = None,
+                 minimum_in_sync_replicas: Optional[pulumi.Input[int]] = None,
+                 partitions: Optional[pulumi.Input[int]] = None,
                  project: Optional[pulumi.Input[str]] = None,
-                 replication: Optional[pulumi.Input[float]] = None,
-                 retention_bytes: Optional[pulumi.Input[float]] = None,
-                 retention_hours: Optional[pulumi.Input[float]] = None,
+                 replication: Optional[pulumi.Input[int]] = None,
+                 retention_bytes: Optional[pulumi.Input[int]] = None,
+                 retention_hours: Optional[pulumi.Input[int]] = None,
                  service_name: Optional[pulumi.Input[str]] = None,
                  termination_protection: Optional[pulumi.Input[bool]] = None,
                  topic_name: Optional[pulumi.Input[str]] = None,
@@ -55,15 +55,15 @@ class KafkaTopic(pulumi.CustomResource):
         :param str resource_name: The name of the resource.
         :param pulumi.ResourceOptions opts: Options for the resource.
         :param pulumi.Input[str] cleanup_policy: Topic cleanup policy. Allowed values: delete, compact.
-        :param pulumi.Input[float] minimum_in_sync_replicas: Minimum required nodes in-sync replicas (ISR) to produce to a partition.
-        :param pulumi.Input[float] partitions: Number of partitions to create in the topic.
+        :param pulumi.Input[int] minimum_in_sync_replicas: Minimum required nodes in-sync replicas (ISR) to produce to a partition.
+        :param pulumi.Input[int] partitions: Number of partitions to create in the topic.
         :param pulumi.Input[str] project: and `service_name` - (Required) define the project and service the topic belongs to.
                They should be defined using reference as shown above to set up dependencies correctly.
                These properties cannot be changed once the service is created. Doing so will result in
                the topic being deleted and new one created instead.
-        :param pulumi.Input[float] replication: Replication factor for the topic.
-        :param pulumi.Input[float] retention_bytes: Retention bytes.
-        :param pulumi.Input[float] retention_hours: Retention period in hours, if -1 it is infinite.
+        :param pulumi.Input[int] replication: Replication factor for the topic.
+        :param pulumi.Input[int] retention_bytes: Retention bytes.
+        :param pulumi.Input[int] retention_hours: Retention period in hours, if -1 it is infinite.
         :param pulumi.Input[str] service_name: Service to link the kafka topic to
         :param pulumi.Input[bool] termination_protection: It is a Terraform client-side deletion protection, which prevents a Kafka topic from being deleted. It is recommended to
                enable this for any production Kafka topic containing critical data.
@@ -119,12 +119,12 @@ class KafkaTopic(pulumi.CustomResource):
             id: pulumi.Input[str],
             opts: Optional[pulumi.ResourceOptions] = None,
             cleanup_policy: Optional[pulumi.Input[str]] = None,
-            minimum_in_sync_replicas: Optional[pulumi.Input[float]] = None,
-            partitions: Optional[pulumi.Input[float]] = None,
+            minimum_in_sync_replicas: Optional[pulumi.Input[int]] = None,
+            partitions: Optional[pulumi.Input[int]] = None,
             project: Optional[pulumi.Input[str]] = None,
-            replication: Optional[pulumi.Input[float]] = None,
-            retention_bytes: Optional[pulumi.Input[float]] = None,
-            retention_hours: Optional[pulumi.Input[float]] = None,
+            replication: Optional[pulumi.Input[int]] = None,
+            retention_bytes: Optional[pulumi.Input[int]] = None,
+            retention_hours: Optional[pulumi.Input[int]] = None,
             service_name: Optional[pulumi.Input[str]] = None,
             termination_protection: Optional[pulumi.Input[bool]] = None,
             topic_name: Optional[pulumi.Input[str]] = None) -> 'KafkaTopic':
@@ -136,15 +136,15 @@ class KafkaTopic(pulumi.CustomResource):
         :param pulumi.Input[str] id: The unique provider ID of the resource to lookup.
         :param pulumi.ResourceOptions opts: Options for the resource.
         :param pulumi.Input[str] cleanup_policy: Topic cleanup policy. Allowed values: delete, compact.
-        :param pulumi.Input[float] minimum_in_sync_replicas: Minimum required nodes in-sync replicas (ISR) to produce to a partition.
-        :param pulumi.Input[float] partitions: Number of partitions to create in the topic.
+        :param pulumi.Input[int] minimum_in_sync_replicas: Minimum required nodes in-sync replicas (ISR) to produce to a partition.
+        :param pulumi.Input[int] partitions: Number of partitions to create in the topic.
         :param pulumi.Input[str] project: and `service_name` - (Required) define the project and service the topic belongs to.
                They should be defined using reference as shown above to set up dependencies correctly.
                These properties cannot be changed once the service is created. Doing so will result in
                the topic being deleted and new one created instead.
-        :param pulumi.Input[float] replication: Replication factor for the topic.
-        :param pulumi.Input[float] retention_bytes: Retention bytes.
-        :param pulumi.Input[float] retention_hours: Retention period in hours, if -1 it is infinite.
+        :param pulumi.Input[int] replication: Replication factor for the topic.
+        :param pulumi.Input[int] retention_bytes: Retention bytes.
+        :param pulumi.Input[int] retention_hours: Retention period in hours, if -1 it is infinite.
         :param pulumi.Input[str] service_name: Service to link the kafka topic to
         :param pulumi.Input[bool] termination_protection: It is a Terraform client-side deletion protection, which prevents a Kafka topic from being deleted. It is recommended to
                enable this for any production Kafka topic containing critical data.
@@ -178,7 +178,7 @@ class KafkaTopic(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="minimumInSyncReplicas")
-    def minimum_in_sync_replicas(self) -> pulumi.Output[Optional[float]]:
+    def minimum_in_sync_replicas(self) -> pulumi.Output[Optional[int]]:
         """
         Minimum required nodes in-sync replicas (ISR) to produce to a partition.
         """
@@ -186,7 +186,7 @@ class KafkaTopic(pulumi.CustomResource):
 
     @property
     @pulumi.getter
-    def partitions(self) -> pulumi.Output[float]:
+    def partitions(self) -> pulumi.Output[int]:
         """
         Number of partitions to create in the topic.
         """
@@ -205,7 +205,7 @@ class KafkaTopic(pulumi.CustomResource):
 
     @property
     @pulumi.getter
-    def replication(self) -> pulumi.Output[float]:
+    def replication(self) -> pulumi.Output[int]:
         """
         Replication factor for the topic.
         """
@@ -213,7 +213,7 @@ class KafkaTopic(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="retentionBytes")
-    def retention_bytes(self) -> pulumi.Output[Optional[float]]:
+    def retention_bytes(self) -> pulumi.Output[Optional[int]]:
         """
         Retention bytes.
         """
@@ -221,7 +221,7 @@ class KafkaTopic(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="retentionHours")
-    def retention_hours(self) -> pulumi.Output[Optional[float]]:
+    def retention_hours(self) -> pulumi.Output[Optional[int]]:
         """
         Retention period in hours, if -1 it is infinite.
         """
